// src/components/models/Reit.tsx
export interface Reit {
  companyid: number
  companyname: string
  ticker: string
  price: number
  sectorid: number
  sectorname: string
  subsectorid: number
  subsectorname: string
  segment: string
  segmentid: number
  gestao: number
  gestao_f: string
  dy: number
  p_vp: number
  valorpatrimonialcota: number
  liquidezmediadiaria: number
  percentualcaixa: number
  dividend_cagr: number
  cota_cagr: number
  numerocotistas: number
  numerocotas: number
  patrimonio: number
  lastdividend: number
}